+++ foundation_stdlib/tests/iodata_test.lua
@@ -4,6 +4,16 @@ local m = foundation.com
 local case = Luna:new("foundation.com.iodata_*")
 
 case:describe("iodata_to_string/1", function (t2)
+  t2:test("can convert a table to a string", function (t3)
+    local result = m.iodata_to_string({"Hello", ", ", "World"})
+    t3:assert_eq(result, "Hello, World")
+  end)
+
+  t2:test("can handle nested tables", function (t3)
+    local result = m.iodata_to_string({"(", {"24", ",", "26", ",", "01"}, ")"})
+    t3:assert_eq(result, "(24,26,01)")
+  end)
+
   t2:test("can return a string given a string", function (t3)
     t3:assert_eq("abc", m.iodata_to_string("abc"))
   end)