+++ foundation_stdlib/tests/value_test.lua
@@ -3,6 +3,26 @@ local m = foundation.com
 
 local case = Luna:new("foundation.com.value")
 
+case:describe("is_blank/1", function (t2)
+  t2:test("nil is blank", function (t3)
+    t3:assert(m.is_blank(nil))
+  end)
+
+  t2:test("an empty string is blank", function (t3)
+    t3:assert(m.is_blank(""))
+  end)
+
+  t2:test("a string with only whitespaces is not blank, though it should be", function (t3)
+    t3:refute(m.is_blank(" "))
+    t3:refute(m.is_blank("    "))
+  end)
+
+  t2:test("booleans are not blank", function (t3)
+    t3:refute(m.is_blank(false))
+    t3:refute(m.is_blank(true))
+  end)
+end)
+
 case:describe("deep_equals/2", function (t2)
   t2:test("can compare 2 scalar values", function (t3)
     t3:assert(m.deep_equals(0, 0))